q-inter: fix bug empty boxes in the list common.mak: remove dependency to an interval lib when compiling with another makefile: add missing makefile in numeric/ and combinatorial/ TestInterval: add test proj_mul

+++ src/contractor/ibex_CtcPropag.cpp
@@ -124,6 +124,6 @@ void  CtcPropag::contract(IntervalVector& _box, const Indicators& indic) {
 
 }
 
-const REAL CtcPropag::default_ratio = __IBEX_DEFAULT_RATIO_PROPAG;
+const double CtcPropag::default_ratio = __IBEX_DEFAULT_RATIO_PROPAG;
 
 } // namespace ibex
